+++ covenant-signer/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/babylonlabs-io/covenant-emulator/covenant-signer/cmd"
+
+func main() {
+	_ = cmd.Execute()
+}